Arbiters got s-m-r-t Smart
Addded a Simple() fucntion that does basic comparisons of binary blobs,
which will be necessary for purely binary streams.  The guts of the old
waitForResponse was reworked into readUntil which takes a function
pointer argument that does all the hard work.  It makes the Control and
Simple nearly identical from a higher level heuristic